put all the python functions in python/processorlib.py and modified daqprocessor Signed-off-by: acolijn <auke.pieter.colijn@gmail.com>
